Add files via upload
This is just a collection of some of the works I had to do in classes or as hands-on project. The main purpose is to show my use of code. I am aware that in some of them, since  only the coding was required at the time, I did demonstrate concern about other readers. That involves random selection of variable names and lack of comments and explanations to readers. Here most of the work is in R, also with some in Python or Tableau.